Created UI for task input
For now , I have chosen to show every element of the task structure in
the UI, mostly for development convenience. But, in the future, I may
hide some of the elements that most users aren't going to care about
(such as priority, or status). That'll most likely be done using a
fold-out menu, but I'll worry about that another time.

I've also created a separate file for custom UI elements which I will
commonly use, and created a custom element for a percentage slider.
Hopefully this file comes in handy even more later down the line, but
it's already improving development convenience.