Created a topbar
The topbar sits at the top of the task list, and contains utilities for
working with the list, such as sorting, or whether to show tasks that
have been completed. More functionality will probably be added to this
later down the line.

Also, not all fields of the task are shown in the list, so that will
have to be worked on, as well as the fact that a due date can't
currently be set for tasks.