Changes to Task structure
Some of the fields of the Task structure didn't properly match their
iCal VTODO equivalents. For example, the priority field isn't actually a percentage,
it's just an integer from 0-10. Also, the Status field isn't actually an
arbitrary string, there is a set of values it can have, and if it isn't
one of those, the todo item is invalid.

I have changed those fields in the struct to better represent their
VTODO counterparts.