Added parser from TaskList to file
Check out parser.rs for implementation details, it's all fully
commented. You can test it out by downloading this commit and running `cargo
run` in the root of the project.

I also forgot to change the sorting implementation for the
due date field when I changed it to an Option, so that's fixed now.

The Task struct also now has a field for storing when the task was
created, which defaults to today, and is also read from any imported ics
files as well as written to any exported ics files.