Made parse from file function safer
Before, if the parse function encountered a file that was invalid in any
way, whether that was a file that doesn't exist, one that contains a field that
wasn't formatted correctly, or a todo item that wasn't properly closed
when it was opened, then the program would just panic and close. Now
I've replaced every call to a function that panics with code that returns
an error that describes what went wrong. This way whatever the function
is called, the code which calls it can take that error value and act
accordingly.