Refactor of parsing code
I only just finished the parsing code, but there were some changes I
wanted to make that I only really thought of after I was already done.
This commit implements those changes, so here's a list of them:

- All functions in parser.rs now have proper doc-tests - the
  TaskList::from_file function had some testing already, but it wasn't
  ideal - the assertion checked the string representation of a TaskList
  instead of checking the data itself, so I changed that. I also wanted
  full test coverage, so I added tests to the other two functions in the file.
- The function to parse from a TaskList to a file now returns a String
  containing said file's potential contents instead of creating a file
  with those contents. This just makes the system more flexible and
  allows for whatever code is implementing the function to choose how to
  deal with the data.
- Some other behind the scenes optimizations have been made, such as:
    - The function to parse from a file to a TaskList no longer uses an
      outdated library to convert from a hex color to an rgb color.
    - The function to convert from a chrono::NaiveDateTime to an ics
      date-time string no longer uses rust string manipulation, it now
      uses chrono::NaiveDateTime's `format()` method. this makes it much
      simpler and (in my opinion) easier to understand.

A note on the tests - I understand that in the design document, I said I
would use test-driven development whenever it applied, and this would
have been a perfect time to use it. The reason I didn't is because I
didn't realize that it applied in this situation until I had already
implemented the functionality. On the plus side, now that I've put the
tests in, I can now be sure the functions work as intended as long as the
tests pass.