Gave `Task`s a UUID field
The need to do this arose when I found a bug that if there are multiple
tasks which use the status field in the input task list, then they have
the same ID, because an egui::ComboBox generates its ID from the label
it's given (which is really stupid). This confuses egui, and so each
ComboBox needs to be spawned with a unique ID.
At first I tried spawning it with a UUID generated with the `uuid`
crate, but that makes it spawn a new UUID every time the UI updates,
which makes the ComboBox unable to be interacted with. So, the ComboBox
needs to be spawned with a UUID that ISN'T generated at the time of
rendering.

The only way I could think of to do this is to add a UUID to
the Task struct, which is a bit of a pain, but I think it'll be useful
later anyway. I also used it for the parser function from TaskList to
file, so it's already got a use I guess.