Added separators between tasks in classic view
Hey, look, another use for a task's UUID! To implement this, when the
classic view begins rendering, I cache the UUID of the final task in the
list, and check that against the UUID of the task that is addressed in
the `filter` closure. If that check determines that it's not the last
class in the list, it draws a separator. Neat!

I also realized that because I added a new field to the Task struct and
didn't change the tests (namely the parse from file to TaskList test)
accordingly, they didn't actually run properly, and so didn't pass, even
though the function itself hasn't changed. So that's fixed now.