Added UUID parsing to parse from file function
Since iCal VTODO items have a uid field anyway, I thought I might as
well try to parse it in the event that it follows the traditional UUID
format. That's not always the case though, so if it isn't a valid UUID
then the task just keeps the ID that's generated using the
Task::default() function.