Improved task editing modal
Yet another use case for task UUIDs! Task modals now use the built-in
egui window titlebar, which allows the windows to be more flexible and
generally more idiomatic in terms of egui code, but the bigger change
is that multiple tasks can now have their modals opened at once, so you
can edit multiple tasks at the same time, which is a huge boon for
workflow versatility.

The reason UUIDs were useful here is that egui windows are yet another
UI element which uses the label it is spawned with as its id. This means
that if any two windows are spawned using the same name, then they use
the same ID and weird stuff happens. So I had to restrict the code to
only allow opening one window at once, so that overlap never happened.
But with UUIDs, different tasks are able to have the same name because
they have unique IDs, so none of that weirdness happens!